IMPALA-2284: Disallow long (1<<30) strings in group_concat()
This is the first step to fix issues with large memory allocations. In
this patch, the built-in `group_concat` is no longer allowed to allocate
arbitraryly large strings and crash impala, but is limited to the upper
bound of possible allocations in Impala.

This patch does not perform any functional change, but rather avoids
unnecessary crashes. However, it changes the parameter type of
FindChunk() in MemPool to be a signed 64bit integer. This change allows
the mempool to allocate internally memory of more than one 1GB, but the
public interface of Allocate() is not changed, so the general limitation
remains. The reason for this change is as follows:

  1) In a UDF FunctionContext::Reallocate() would allocate slightly more
  than 512MB from the FreePool.
  2) The free pool tries to double this size to alloocate 1GB from the
  MemPool.
  3) The MemPool doubles the size again and overflows the signed 32bit
  integer in the FindChunk() method. This will then only allocate 1GB
  instead of the expected 2GB.

What happens is that one of the callers expected a larger allocation
than actually happened, which will in turn lead to memory corruption as
soon as the memory is accessed.

Reviewed-on: http://gerrit.cloudera.org:8080/858
Tested-by: Internal Jenkins

+++ be/src/runtime/mem-pool.cc
@@ -45,7 +45,7 @@ MemPool::MemPool(MemTracker* mem_tracker, int chunk_size)
   DCHECK(mem_tracker != NULL);
 }
 
-MemPool::ChunkInfo::ChunkInfo(int size)
+MemPool::ChunkInfo::ChunkInfo(int64_t size)
   : owns_data(true),
     data(reinterpret_cast<uint8_t*>(malloc(size))),
     size(size),
@@ -89,7 +89,7 @@ void MemPool::FreeAll() {
   }
 }
 
-bool MemPool::FindChunk(int min_size, bool check_limits) {
+bool MemPool::FindChunk(int64_t min_size, bool check_limits) {
   // Try to allocate from a free chunk. The first free chunk, if any, will be immediately
   // after the current chunk.
   int first_free_idx = current_chunk_idx_ + 1;
@@ -111,7 +111,7 @@ bool MemPool::FindChunk(int min_size, bool check_limits) {
 
   if (current_chunk_idx_ == static_cast<int>(chunks_.size())) {
     // need to allocate new chunk.
-    int chunk_size = chunk_size_;
+    int64_t chunk_size = chunk_size_;
     if (chunk_size == 0) {
       if (current_chunk_idx_ == 0) {
         chunk_size = DEFAULT_INITIAL_CHUNK_SIZE;
